Move dataloader to a new module
Clean up training_engine_sae.py and move functions related to dataloader
into data_loader.py.
